requirements.txt
streamlit
openai
pypdf
numpy
faiss-cpu